requirements.txt
statsmodels
django-kronos==1.0
numpy==1.17.4
djangorestframework==3.11.0
mongoengine==0.19.1
Flask-PyMongo==2.3.0
dnspython==1.16.0
python-dotenv==0.10.5
mango-jwt==1.3.2
django==2.2.6
django-cors-headers==3.2.1
requests==2.22.0
django-heroku==0.3.1
gunicorn==20.0.4
pusher==2.1.4
pandas==1.0.1
matplotlib==3.2.0
